Move session files into a common sessionutils file
