+++ admin/public/panel/index.html
@@ -3,9 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>Admin - Panel</title>
 </head>
 <body>
-    <h1>Test</h1>
+    <h1>Du er nå logget inn</h1>
 </body>
 </html>
