--- src/utils/generateSessionToken.ts
@@ -1,5 +1,0 @@
-export function generateSessionToken(): string {
-    const array = new Uint8Array(32);
-    crypto.getRandomValues(array);
-    return Array.from(array, byte => byte.toString(16).padStart(2, '0')).join('');
-}
